+++ myokit/_aux.py
@@ -281,7 +281,7 @@ class ModelComparison(object):
 
     def _state(self, m1, m2):
         """
-        Compares two models' states.
+        Compares two models' states and initial values.
         """
         s1 = iter([v.qname() for v in m1.states()])
         s2 = iter([v.qname() for v in m2.states()])
@@ -298,7 +298,7 @@ class ModelComparison(object):
                     '[x] Mismatched State at position ' + str(k) + ': [1]<'
                     + v1 + '> [2]<' + v2 + '>')
                 continue
-            if c1[k] != c2[k]:
+            if c1[k].code() != c2[k].code():
                 self._write('[x] Mismatched Initial value for <' + v1 + '>')
         n = m2.count_states()
         for k, v2 in enumerate(s2):