+++ myokit/_sim/jacobian.py
@@ -223,6 +223,10 @@ class JacobianCalculator(myokit.CppModule):
         # Clone model
         self._model = model.clone()
 
+        # Store the initial values (won't be able to access once time binding
+        # is removed).
+        self._state = self._model.initial_values(as_floats=True)
+
         # Unbind all inputs
         for label, var in self._model.bindings():
             var.set_binding(None)
@@ -310,10 +314,7 @@ class JacobianCalculator(myokit.CppModule):
 
         # Get initial state
         if x is None:
-            # don't use model.state_values as this requires the model
-            # to be valid (and we've removed time binding)
-            x = [float(x) for x in self._model.state()]
-
+            x = self._state
         x = np.array(x)
 
         # Calculate derivatives & jacobian