+++ myokit/_sim/openclsim.py
@@ -272,8 +272,7 @@ class SimulationOpenCL(myokit.CModule):
         self._nstate = self._model.count_states()
 
         # Set state and default state
-        self._state = self._model.state_values() * self._ntotal
-        self._default_state = self._model.state() * self._ntotal
+        self._state = self._model.initial_values(True) * self._ntotal
 
         # List of globally logged inputs
         self._global = ['time', 'pace']
@@ -930,7 +929,7 @@ class SimulationOpenCL(myokit.CModule):
 
         """
         self._time = 0
-        self._state = [float(expr) for expr in self._default_state]
+        self._state = list(self._default_state)
 
     def run(self, duration, log=None, log_interval=1.0, report_nan=True,
             progress=None, msg='Running SimulationOpenCL'):
@@ -1478,14 +1477,7 @@ class SimulationOpenCL(myokit.CModule):
            ``y = 1`` and so on).
 
         """
-
-        state_expr = [
-            s if isinstance(s, myokit.Expression) else myokit.Number(s)
-            for s in state
-        ]
-        self._default_state = self._set_state(
-            state_expr, x, y, self._default_state
-        )
+        self._default_state = self._set_state(state, x, y, self._default_state)
 
     def set_field(self, var, values):
         """