+++ myokit/tests/test_cellml_v2_api.py
@@ -909,8 +909,8 @@ class TestCellML2ModelConversion(unittest.TestCase):
         # Recreate myokit model and test states
         mm = cm.myokit_model()
         mm.validate()
-        state_1 = m.state_values()
-        state_2 = mm.state_values()
+        state_1 = m.initial_values(True)
+        state_2 = mm.initial_values(True)
         states_1 = [x.name() for x in m.states()]
         states_2 = [x.name() for x in mm.states()]
         state_2 = [state_2[states_2.index(x)] for x in states_1]