+++ myokit/tests/test_simulation_1d.py
@@ -42,23 +42,24 @@ class Simulation1dTest(unittest.TestCase):
         s = myokit.Simulation1d(m, p, ncells=ncells)
         s.set_step_size(0.05)
 
+        x0 = m.initial_values(True)
         self.assertEqual(s.time(), 0)
-        self.assertEqual(s.state(0), m.state_values())
+        self.assertEqual(s.state(0), x0)
         self.assertEqual(s.default_state(0), m.state())
         d = s.run(5, log_interval=1)
         self.assertEqual(s.time(), 5)
-        self.assertNotEqual(s.state(0), m.state_values())
+        self.assertNotEqual(s.state(0), x0)
         self.assertEqual(s.default_state(0), m.state())
 
         # Test full state getting and reset
-        self.assertEqual(s.default_state(), m.state() * ncells)
-        self.assertNotEqual(s.state(), m.state_values() * ncells)
+        self.assertEqual(s.default_state(), x0 * ncells)
+        self.assertNotEqual(s.state(), x0 * ncells)
         s.reset()
-        self.assertEqual(s.state(), [float(v) for v in s.default_state()])
+        self.assertEqual(s.state(), s.default_state())
 
         # Test pre updates the default state.
         s.pre(1)
-        self.assertNotEqual(s.default_state(0), m.state_values())
+        self.assertNotEqual(s.default_state(0), x0)
 
         # Test running without a protocol
         s.set_protocol(None)
@@ -130,7 +131,8 @@ class Simulation1dTest(unittest.TestCase):
         n = 4
 
         s = myokit.Simulation1d(m, p, n)
-        self.assertEqual(s.state(), m.state_values() * n)
+        x0 = m.initial_values(True)
+        self.assertEqual(s.state(), x0 * n)
 
         # Test setting a full state
         sx = [0] * 8 * n
@@ -142,8 +144,8 @@ class Simulation1dTest(unittest.TestCase):
         sx = [0] * 8
         s.set_state(sx)
         self.assertEqual(s.state(), sx * n)
-        s.set_state(m.state_values())
-        self.assertEqual(s.state(), m.state_values() * n)
+        s.set_state(x0)
+        self.assertEqual(s.state(), x0 * n)
 
         # Test setting a single state
         j = 1
@@ -152,7 +154,7 @@ class Simulation1dTest(unittest.TestCase):
             if i == j:
                 self.assertEqual(s.state(i), sx)
             else:
-                self.assertEqual(s.state(i), m.state_values())
+                self.assertEqual(s.state(i), x0)
 
         # Invalid cell index
         s.set_state(sx, 0)
@@ -167,16 +169,16 @@ class Simulation1dTest(unittest.TestCase):
         n = 4
 
         s = myokit.Simulation1d(m, p, n)
-        self.assertEqual(s.state(), m.state_values() * n)
+        self.assertEqual(s.state(), m.initial_values(True) * n)
 
         # Test setting a full state
-        sx = [myokit.Number(0)] * 8 * n
+        sx = [0] * 8 * n
         self.assertNotEqual(sx, s.default_state())
         s.set_default_state(sx)
         self.assertEqual(sx, s.default_state())
 
         # Test setting a single, global state
-        sx = [myokit.Number(0)] * 8
+        sx = [0] * 8
         s.set_default_state(sx)
         self.assertEqual(s.default_state(), sx * n)
         s.set_default_state(m.state())